# local_eat/api/docs/swagger.yaml
basePath: /
definitions:
  model.Producers:
    properties:
      created:
        example: "2020-01-01"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John
        type: string
      picture:
        example: John.jpg
        type: string
    type: object
  model.Users:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: "1234"
        type: string
      username:
        example: John
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server local eat API server.
  title: local eat API
  version: "1.0"
paths:
  /api/auth/authenticate:
    get:
      description: Validate user token
      produces:
      - application/json
      responses:
        "200":
          description: User authenticated
        "401":
          description: Unauthorized
      summary: Validate user token
      tags:
      - Auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Send username and password to login to receive a token in a cookie
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.Users'
      responses:
        "200":
          description: User authenticated
        "400":
          description: Invalid password
        "500":
          description: Internal server error
      summary: Send username and password to login
      tags:
      - Auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Send user data to create a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.Users'
      responses:
        "200":
          description: User created
        "400":
          description: Invalid request
        "500":
          description: Internal server error
      summary: Send user data to create a new user
      tags:
      - Auth
  /api/producers:
    get:
      description: Get producers id, name, picture and created values
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Producers'
            type: array
        "404":
          description: Not found
        "500":
          description: Internal server error
      summary: Get producers
      tags:
      - Producers
produces:
- application/json
schemes:
- http
swagger: "2.0"